Got some movement, but not perfect. Gutted float-based duty cycle in leiu of standard i2c input.

+++ models/src/lib.rs
@@ -23,8 +23,8 @@ pub enum MotorName{
 #[derive(PartialEq)]
 pub struct MotorData{
     pub motor_name: MotorName,
-    pub max: f64,
-    pub min: f64
+    pub max: u16,
+    pub min: u16
 }
